feat(tofs): lookup files directory in “tpldir” hierarchy
Without user configuration, the “libtofs.jinja” lookup the “files”
directory under “tplroot” with is the base directory of the current
formula unless the “v1_path_prefix” is set.

Now, we lookup the directory in each sub-directory from the
sub-directory of the current SLS (“tpldir”) to the top level directory
of the formula (“tplroot”).

This permit to support formula with different components, each one
providing its own “files” directory like “systemd-formula”.