feat(map.jinja): load configuration values from configurable sources
Lookup the configuration for `map.jinja` from:

1. builtin default

   1. `osarch` grain
   2. `os_family` grain
   3. `os` grain
   4. `osfinger` grain
   5. `lookup` table retrived by `config.get`
   6. configuration retrived by `config.get`
   7. `id` grain

2. `defaults.yaml`: optionally define a formula specific
   `map_jinja:sources`

3. global configuration lookup `map_jinja:sources`

4. formula specific `<tplroot>:map_jinja:sources`

The hard coded default is backward compatible and add the minion id at
the end of the list.

If an entry does not match a `salt['config.get']` parameter, it's used
as a literal file path (with `.yaml` added if it's missing).

Each YAML file are formatted with the following top level keys:

- `values`: contains the the parameter values

- `strategy` (optional): define the merge strategy for the function
  `salt.slsutils.merge` (aggregate, list, overwrite, recurse,
  smart). The default is `smart`.

- `merge_lists` (optional): boolean to merge lists or overwrite them

It's possile to configure `config.get` when looking up the formula
configuration. You need to define a subkey `strategy` with the
following differences from YAML files:

- `strategy` can only be one of `None`, `overwrite` and `recurse`.
  The default is `None`.

- if you use `salt-ssh`, this merge strategy is skipped and and error
  log message is emitted.

The parameters values are merged in the following order:

1. initialize default values from `defaults.yaml`

2. merge the values from each source defined by the ordered list
   `map_jinja:sources`